TasksPage: extract getPreviewEnv() into an external helper file
